refact(api): migrate local BaseModel subclasses — code domain (#6042)
Phase 2 of #6042: migrate schemas from 5 code-domain endpoint files into schemas_code.py.

filesystem_mcp.py (27 classes) → schemas_code.py
  - 14 response classes (FilesystemReadTextResponse … FilesystemListAllowedResponse,
    FilesystemSecurityInfo)
  - MCPTool (shared with git_mcp — single definition for all MCP tool lists)
  - 12 request classes (ReadTextFileRequest … GetFileInfoRequest)

git_mcp.py — removes local duplicate MCPTool, imports shared MCPTool from schemas_code
  (6 request models deferred: reference module-level constants/regex from git_mcp.py)

code_intelligence.py (7 classes) → schemas_code.py
  - SuggestionItem, SuggestionsResponse, FindingsPlaceholderResponse
  - TaskStatusResponse → CodeIntelligenceTaskStatusResponse (avoids conflict with settings.py)
  - StartTaskResponse, ClearStuckResponse, CachedSecurityScoreResponse

playwright.py (5 classes) → schemas_code.py
  - SearchRequest → PlaywrightSearchRequest
  - ScreenshotRequest → PlaywrightScreenshotRequest
  - NavigateRequest → PlaywrightNavigateRequest
  - ReloadRequest → PlaywrightReloadRequest
  - InteractRequest → PlaywrightInteractRequest
  (TestMessageRequest/FrontendTestRequest deferred: depend on runtime ssot_config)

ide_integration.py (8 classes) → schemas_code.py
  - IDERulesResponse, IDEConfigUpdateResponse, IDECategoriesResponse,
    IDESeveritiesResponse, IDEBatchAnalyzeResponse, IDEHealthResponse,
    IDECategoryItem, IDESeverityItem
  (Generic LSP types Position/Range/Diagnostic etc. deferred to follow-up PR: 62 internal
  usages in IDEIntegrationEngine class require careful rename pass)
